+++ native/c/zut.hpp
@@ -13,10 +13,9 @@
 #define ZUT_HPP
 
 #include <iconv.h>
-#include <iostream>
 #include <vector>
 #include <string>
-#include "zcntype.h"
+#include "ztype.h"
 
 /**
  * @struct ZConvData
@@ -220,5 +219,11 @@ void zut_debug_message(const char *message);
  * @return True if a should come before b in sorted order, false otherwise
  */
 bool zut_string_compare_c(const std::string &a, const std::string &b);
+
+/**
+ * @brief List a parmlib
+ * @return Return code (0 for success, non-zero for error)
+ */
+int zut_list_parmlib(ZDIAG &diag, std::vector<std::string> &parmlibs);
 
 #endif // ZUT_HPP